pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "grainrx"
version = "1.0.0"
description = "Physics-based film grain synthesis using the inhomogeneous Boolean model"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "numba>=0.58.0",
    "Pillow>=10.0.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
ui = ["fastapi>=0.100.0", "uvicorn>=0.23.0"]

[project.scripts]
grainrx = "core.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]